BLGMIGRATION-866 clean: refactor
